Add simple TCP server to execute commands
Create a simple TCP server that executes arbitrary shell commands sent
to it by a client, i.e. telnet or netcat, and returns the ouput back to
the client.

Commands are separated by newlines; I'm using the [net/textproto][]
package, which is designed to help implement text-based protocols, to
parse the input.

Connections can be closed and new ones reopened as desired. The server
will remain running indefinitely.

[net/textproto][]: https://golang.org/pkg/net/textproto/