Hook strrchr() rather than __libc_start_main
I wasn't able to hook __libc_start_main as `dl.Sym()` was rejecting the
function signature.

Instead, hook `strrchr()` because it has a much simpler function
signature and is called immediately when `ps` or `top` are invoked.
There's probably a better function to hook, but this is enough to
get this demonstration working.

Run `backdoor()` in a goroutine to prevent it from blocking the call we
make to the original `strrchr()` function we are wrapping.

For more details of how we're calling the original function we're
wrapping/hooking, see:
http://opensourceforu.efytimes.com/2011/08/lets-hook-a-library-function/
f